requirements.txt
langchain
openai
python-decouple
pypdf
tiktoken
faiss-cpu
boto3
streamlit
streamlit-chat
